Act008_Main.java Act008_Main_Presenter.java Act008_Main_Presenter_Impl.java     - Navegação apra nova tela de ações do serial. @#$ Act083_Main_Presenter.kt     - Tratamento de null. @#$ Act092_Main.kt Act092Presenter.kt ActionSerialRepository.kt ActionUseCases.kt IActionSerialRepository.kt ToolBox_Inf.java     - Adição de usecase para chamada de serviço de ações não focadas. @#$ AppBase.java ConstantBaseApp.java     - Configuração de path para arquivo json de não focados. @#$ Constant.java     - correção na url do serviço de não focadaos; @#$ MyActionsCache.kt TUnfocusAndHistoricRec.kt    - Correção no cabeçalho da response do serviço de não focados. @#$ TUnfocusAndHistoricEnv.kt     -Ajuste de tipo. @#$ WS_UnfocusAndHistoric.kt     - Adição de save do arquivo de outras ações. @#$

+++ app/src/main/java/com/namoadigital/prj001/model/TUnfocusAndHistoricEnv.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 class TUnfocusAndHistoricEnv(
     @SerializedName("product_code") val productCode: Int,
-    @SerializedName("serial_code") val serialCode: Int,
+    @SerializedName("serial_code") val serialCode: Long,
     appCode: String,
     appVersion: String?,
     appType: String?,
